feat(thumbnail): Thumbnail generator as a service
