+++ src/jdRoll/service/persoService.php
@@ -14,9 +14,10 @@ class PersoService {
     private $db;
     private $session;
 
-    public function __construct($db, $session) {
+    public function __construct($db, $session, $thumbnailService) {
         $this->db = $db;
         $this->session = $session;
+        $this->thumbnailService = $thumbnailService;
     }
 
 
@@ -197,7 +198,7 @@ class PersoService {
         $stmt->execute();
 
         //Generate thumbnail from avatar
-        $this->_generateThumbnail($perso_id, $request->get('avatar'));
+        $this->thumbnailService->generateThumbnail('perso', $perso_id, $request->get('avatar'));
     }
 
     public function setTechnical($perso_id, $template) {
@@ -252,7 +253,7 @@ class PersoService {
         $stmt->execute();
 
         //Generate thumbnail
-        $this->_generateThumbnail($this->db->lastInsertId(), $request->get('avatar'));
+        $this->thumbnailService->generateThumbnail('perso', $this->db->lastInsertId(), $request->get('avatar'));
     }
 
     public function deletePersonnage($perso_id) {
@@ -389,160 +390,6 @@ class PersoService {
 		campagne_id = :id";
         $result = $this->db->fetchAll($sql, array("id" => $campagne_id));
         return $result;
-    }
-
-
-    /**
-     * Generate thumbnail for all perso
-     */
-    public function generateThumbnails(){
-        //Size of the thumbnail
-        $sql = "SELECT id, avatar FROM personnages WHERE 1";
-
-        foreach($this->db->fetchAll($sql) as $perso){
-            $this->_generateThumbnail($perso['id'], $perso['avatar'], $force=false);
-        }
-    }
-
-    /**
-     * Creates a thumbnail for one perso.
-     * if $force=FALSE, will not overwrite existing thumbnail
-     * @param $persoId
-     * @param $avatarPath
-     * @param bool $force
-     */
-    protected function _generateThumbnail($persoId, $avatarPath, $force=true){
-
-        //Square size
-        $size = 64;
-
-        //Create directory if does not exist
-        $folder = FOLDER_FILES . '/thumbnails/';
-        if( ! file_exists($folder)){
-            mkdir(FOLDER_FILES, true);
-            mkdir($folder, true);
-        }
-
-        $isRemote = strpos($avatarPath, 'http') === 0;
-
-        //Compute pathes & extension
-        $thumbPath = $folder . "perso_{$persoId}.png";
-        $filepath = $avatarPath;
-        $extension = null;
-
-        if(file_exists($thumbPath) && !$force){
-            //Thumbnail exists, continue
-            return;
-        }
-        else if( ! $isRemote){
-            //For now, ignore local files as I do not have them
-            $filepath = FOLDER_FILES . '/' . basename($avatarPath);
-            if( ! file_exists($filepath)){
-                $this->_generateDefaultThumbnail($thumbPath);
-            }
-        }
-        else{
-            //Get extension for remote files, sometimes it is not present in the URL and can cause errors
-            $headers = @get_headers($filepath, 1);
-            //If redirection for image, get latest content type
-            $contentType = is_array($headers['Content-Type']) ? $headers['Content-Type'][count($headers['Content-Type'])-1]:$headers['Content-Type'];
-            $extension = str_replace('image/', '', $contentType);
-        }
-
-        //Create thumbnail
-        set_time_limit(30);
-        $result = $this->_makeThumb($filepath, $thumbPath, $size, 100, $extension);
-
-        //If an error arose
-        if($result === false){
-            //Use default empty image
-            $this->_generateDefaultThumbnail($thumbPath);
-        }
-    }
-
-    /**
-     * Shortcut to create a default profile thumbnail
-     * @param $thumbPath
-     */
-    protected function _generateDefaultThumbnail($thumbPath){
-        copy(FOLDER_FILES . '/../img/defaultPerso.png', $thumbPath);
-    }
-
-    /**
-     * Create a squared thumbnail
-     * @param $source
-     * @param $destination
-     * @param int $square_size
-     * @param int $quality
-     * @param $forceExtension
-     * @return bool
-     */
-    protected function _makeThumb($source, $destination, $square_size=167, $quality=90, $forceExtension) {
-
-        $status  = false;
-        list($width, $height, $type, $attr) = @getimagesize($source);
-
-        if( ! $width || ! $height){
-            //Problem loading
-            return false;
-        }
-
-        if($width< $height) {
-            $width_t =  $square_size;
-            $height_t    =   round($height/$width*$square_size);
-            $off_y       =   ceil(($width_t-$height_t)/2);
-            $off_x       =   0;
-
-        } elseif($height< $width) {
-
-            $height_t    =   $square_size;
-            $width_t =   round($width/$height*$square_size);
-            $off_x       =   ceil(($height_t-$width_t)/2);
-            $off_y       =   0;
-
-        } else {
-
-            $width_t    =   $height_t   =   $square_size;
-            $off_x      =   $off_y      =   0;
-        }
-
-        $thumb_p    = imagecreatetruecolor($square_size, $square_size);
-
-        $extension  = $forceExtension ? $forceExtension : pathinfo($source, PATHINFO_EXTENSION);
-
-        if($extension == "gif" or $extension == "png"){
-
-            imagecolortransparent($thumb_p, imagecolorallocatealpha($thumb_p, 0, 0, 0, 127));
-            imagealphablending($thumb_p, false);
-            imagesavealpha($thumb_p, true);
-        }
-
-        if ($extension == 'jpg' || $extension == 'jpeg')
-            $thumb = imagecreatefromjpeg($source);
-        else if ($extension == 'gif')
-            $thumb = imagecreatefromgif($source);
-        else if ($extension == 'png')
-            $thumb = imagecreatefrompng($source);
-        else
-        {echo "Wrong extension: $source ($extension)"; return;}
-
-        $bg = imagecolorallocate ( $thumb_p, 255, 255, 255 );
-        imagefill ($thumb_p, 0, 0, $bg);
-
-        imagecopyresampled($thumb_p, $thumb, $off_x, $off_y, 0, 0, $width_t, $height_t, $width, $height);
-
-        $destinationExtension = pathinfo($destination, PATHINFO_EXTENSION);
-        if ($destinationExtension == 'jpg' || $destinationExtension == 'jpeg')
-            $status = @imagejpeg($thumb_p,$destination,$quality);
-        if ($destinationExtension == 'gif')
-            $status = @imagegif($thumb_p,$destination,$quality);
-        if ($destinationExtension == 'png')
-            $status = @imagepng($thumb_p,$destination,9);
-
-        imagedestroy($thumb);
-        imagedestroy($thumb_p);
-
-        return $status;
     }
 }
